Remove old form + refactor
